added documentation and separated files
