+++ MapCache/Classes/CachedTileOverlayZoomRenderer.swift
@@ -4,34 +4,45 @@
 //
 //  Created by Cameron Deardorff on 9/20/20.
 //
+//
 
 import Foundation
 import MapKit
 
+///
+/// A Tile overlay that supports to zoom
 class CachedTileOverlayZoomRenderer: MKTileOverlayRenderer {
     
+    /// Indicates if the renderer is ready to draw. It´s always true
+    /// - SeeAlso: [MKOverlayRenderer](https://developer.apple.com/documentation/mapkit/mkoverlayrenderer)
     override func canDraw(_ mapRect: MKMapRect, zoomScale: MKZoomScale) -> Bool {
         // very important to call super.canDraw first, some sort of side effect happening which allows this to work (???).
         let _ = super.canDraw(mapRect, zoomScale: zoomScale)
         return true
     }
     
+    /// Draws the tile in the map
+    /// - Parameter mapRect the map rect where the tile has to be drawn
+    /// - Parameter zoomScale current zoom in the map
     override func draw(_ mapRect: MKMapRect, zoomScale: MKZoomScale, in context: CGContext) {
         
+        // use default rendering if the type of overlay is not CachedTileOverlay
         guard let cachedOverlay = overlay as? CachedTileOverlay else {
             super.draw(mapRect, zoomScale: zoomScale, in: context)
             return
         }
+        
         // use default rendering when tiles are available
         guard cachedOverlay.shouldZoom(at: zoomScale) else {
             super.draw(mapRect, zoomScale: zoomScale, in: context)
             return
         }
         
+        //Extract the zoomable tiles for this mapRect
         let tiles = cachedOverlay.tilesInMapRect(rect: mapRect, scale: zoomScale)
+        
         for tile in tiles {
             cachedOverlay.loadTile(at: tile.path) { [weak self] (data, error) in
-                
                 guard let strongSelf = self,
                       let data = data,
                       let provider = CGDataProvider(data: data as CFData),
@@ -51,24 +62,5 @@ class CachedTileOverlayZoomRenderer: MKTileOverlayRenderer {
                 context.restoreGState()
             }
         }
-    }
-}
-
-///
-/// Specifies a single tile and area of the tile that should upscaled
-///
-struct ZoomableTile {
-    let path: MKTileOverlayPath
-    let rect: MKMapRect
-    // delta from given tile z to desired tile z
-    let overZoom: Zoom
-}
-
-extension MKZoomScale {
-    func toZoomLevel(tileSize: CGFloat) -> Int {
-        let numTilesAt1_0 = MKMapSize.world.width / Double(tileSize)
-        let zoomLevelAt1_0 = log2(numTilesAt1_0)
-        let zoomLevel: Double = Double.maximum(0, zoomLevelAt1_0 + Double(floor(log2(self) + 0.5)))
-        return Int(zoomLevel)
     }
 }
